Migrations added / authentication updated.

+++ database/migrations/2026_03_10_230000_add_branch_id_to_users_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
     {
         Schema::table('users', function (Blueprint $table) {
             if (! Schema::hasColumn('users', 'branch_id')) {
-                $table->unsignedBigInteger('branch_id')->nullable()->after('role_id');
+                $table->unsignedInteger('branch_id')->nullable()->after('role_id');
                 $table->foreign('branch_id')->references('branch_id')->on('branches')->onDelete('set null');
             }
         });
